+++ app/src/main/java/woowacourse/movie/presenter/SeatsSelectionPresenter.kt
@@ -3,8 +3,8 @@ package woowacourse.movie.presenter
 import woowacourse.movie.domain.model.MovieTicket
 import woowacourse.movie.domain.policy.PricingPolicy
 import woowacourse.movie.domain.repository.MovieRepository
-import woowacourse.movie.ui.view.booking.SeatsSelectionContract
 import woowacourse.movie.ui.view.seat.SeatButtonState
+import woowacourse.movie.ui.view.seat.SeatsSelectionContract
 
 class SeatsSelectionPresenter(
     private val view: SeatsSelectionContract.View,
